requirements.txt
flask
flask-cors
flask-restplus
flask_httpauth
uwsgi
requests
mecab-python3